modified to change dimension of keypoint

+++ config.py
@@ -131,7 +131,7 @@ class ToTensor(object):
         # numpy image: H x W x C
         # torch image: C X H X W
         image = image.transpose((2, 0, 1))
-        image = torch.from_numpy(image)
-        keypoints = torch.from_numpy(keypoints)
+        image = torch.from_numpy(image).float()
+        keypoints = torch.from_numpy(keypoints).float()
         return {'image': image,
                 'keypoints': keypoints}
